Null checking
I've seen this one throw a null pointer exception a couple of times, when mPopupWindow is null. Not sure how, but I suspect this is one of those typical errors that happen when the activity is closing down or goes out of focus before the window can trigger.

Not sure what one can do to save this otherwise, but suggest being defensive in the code and checking for null before calling it? I figure anything is better than crashing the app with a NullPointerException.