Added support for non-touch screens
Added support for non-touch screens so that they can dismiss tooltips either by pressing the DPAD_CENTER button, or the ENTER button. This is in addition to the BACK button that works by default. This all works only if the tooltip is set to be focusable, so on non-touch devices, we set the tooltip focusable by default.